Rename 'codegen' package to 'engine' package
